first stab at rewriting the activity algo. This is first stage only; comunity profiling coming next

+++ src/main/scala/helix/lib/contracts.scala
@@ -32,6 +32,6 @@ trait Scoring {
   protected def scoring: ScoringStrategy
   
   trait ScoringStrategy {
-    def calculateProjectActivityScore(project: Project): BigInt
+    def calculateProjectActivityScore(project: Project): Double
   }
 }
